+++ src/devtools/components/sidebar/AboutExtension.tsx
@@ -2,6 +2,8 @@ import pkg from "../../../../package.json";
 import styles from "./AboutExtension.module.css";
 
 export default function AboutExtension() {
+  const version = () => (import.meta.env.DEV ? "DEV" : `v${pkg.version}`);
+
   return (
     <div class={styles.about}>
       <a
@@ -9,7 +11,7 @@ export default function AboutExtension() {
         title="https://github.com/ghazi-git/indexeddb-browser/releases"
         target="_blank"
       >
-        v{pkg.version}
+        {version()}
       </a>
       <a
         href="https://github.com/ghazi-git/indexeddb-browser/issues"